Move spi interfaces to cdap-event-writer-spi module
