+++ cdap-event-writer-spi/src/main/java/io/cdap/cdap/spi/events/EventWriterContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2021 Cask Data, Inc.
+ * Copyright © 2022 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,7 +14,20 @@
  * the License.
  */
 
-package io.cdap.cdap.internal.events;
+package io.cdap.cdap.spi.events;
 
-public class EventWriterContext {
+import java.util.Map;
+
+/**
+ * Interface for an {@link EventWriter} context
+ */
+public interface EventWriterContext {
+
+  /**
+   * Returns a map of string key value properties
+   *
+   * @return {@link Map<String, String>}
+   */
+  Map<String, String> getProperties();
+
 }
